requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
xgboost
lightgbm
shap
lime
joblib
streamlit
category_encoders
scipy
python-dotenv
mysql-connector-python
pymysql